+++ changelog.md
@@ -3,8 +3,13 @@
 ## 2023.11.11 - 0.1.0
 - Initial commit 🥳
 
-## 2023.11.14 - 0.2.0
+## 2023.11.13 - 0.2.0
 - Added `trot` binary
 - Added `Actor::input`
 - Added `Response::save_to_path`
 - Added tcp timeout
+
+## 2023.11.13 - 0.3.0
+- Added `--pretty-print` option to `trot`
+- Added `parse` module for parsing gemtext into symbols.
+- Added `Response::is_gemtext`